Bug 11593 - subtag_registry.sql files contain the same information for each language
This patch removes subtag_registry.sql from lang dirs,
and put it on a new mandatory dir

In that way we need to update only one file.
All similar files are removed.

To test:
1) Try a new install, or simply create a new database,
grant permissions on it, change koha-conf.xml to use
this database and go to staff page

2) Go step by step until all mandatory or optional sql
files are loaded. Check that subtag_registry.sql is
loaded.

3) Do it again but on other language. First install
translation and repeat 2), could use de-DE or fr-FR

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described, tested with de-DE.
Thx!

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>