Bug 11944: Fix fund names encoding on the parcel page
On the parcel page, the subtotal values are stored in an hashref. The
key was the fund name. The fund name being encoded in utf8, the hashref
was broken.
This patch replaces this logic creating a key based on an integer.
The budget name is now stored in a new "budget_name" key.

Note: It also fixes a small ergonomic issue (a td was missing in the
tfoot).

Signed-off-by: Paola Rossi <paola.rossi@cineca.it>
Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>
Signed-off-by: Dobrica Pavlinusic <dpavlin@rot13.org>

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>