Bug 12102 - Move humanmsg jQuery plugin outside of language-specific directory
This patch moves the humanmsg jQuery plugin to
intranet-tmpl/prog/lib/jquery/plugins so that it will not be duplicated
for each set of translated templates.

To test, apply the patch and modify a system preference. The
modification messages should appear normally.

Patch works as expected
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Passes all tests and QA script, works as described.
No regressions found.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>