Bug 12195 - Remove duplicated plugin jquery.dataTables.min.js
All templates in the staff client now load the DataTables jquery plugin
through a single include file, datatables.inc, which points to a copy of
the plugin in intranet-tmpl/lib/jquery/plugins. This patch removes the
unused copy of the plugin found in intranet-tmpl/prog/en/lib/jquery/plugins.

To test, search the Koha source for references to
jquery.dataTables.min.js. There should be no files which point to the
redundant file.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>