Bug 12190: fold the "Creators" system preference tab into "Tools"
The "Creators" tab in the system preferences editor had only one
preference.  In order to consolidate the display as well as to
address the issue that "creators" doesn't have much of a connection
to "patron cards", the ImageLimit system preference is moved to
the Tools tab by this patch.

To test:

[1] Verify that the ImageLimit preference now shows up in the
    Tools tab in the system preference editor (along with an
    existing Tools preference, MaxItemsForBatch.
[2] Verify that the sysprefs editor no longer has a "Creators"
    tab.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>