Bug 12205 : Add a system preference allowing hold charges to block circ
Currently, if you have holds charges, they are not taken into
consideration when circulating items.

Manual Invoices, and rental charges are governed by a systempreference

Holds are never counted

And all other charges (overdues, lost items etc) are counted

This patch adds a systempreference to allow Hold charges to be counted
as well.

To test

1/ Set a borrower category to have holds charges
2/ Place a hold for a borrower in that category
3/ Go to checkout, notice that charge is not showing or blocking on
that screen
4/ apply patch
5/ notice that charge now shows on checkout

Signed-off-by: Nick Clemens <nick@quecheelibrary.org>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>