Bug 12308 - show all title subfields on the holds list
The patron summary page in the bootstrap OPAC does not show subtitle in
the list of checkouts, overdues, or holds. This patch corrects this.

To test, set your OPAC theme to bootstrap and log in to the OPAC as a
user who has items checked out, items overdue, and items on hold. At
least one of each of those should be a title which has a subtitle as
defined in Administration -> Keyword to MARC mapping. Subtitles should
appear correctly under the "Checked out," "Overdue," and "Holds" tabs.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>

Pushed to master by me in my role as OPAC module maintainer