Bug 12255 - Amazon cover images have useless image alt text in bootstrap
"item-thumbnail" is not a useful alt text for an Amazon book cover
image. Better would be the link's title text, "View on Amazon.com." This
patch makes this correction. (Rather than "View at Amazon.com" which
doesn't sound correct to me).

To test you must have the OPACAmazonCoverImages system preference
enabled. Log in to the OPAC and view your list of checked-out titles.
The alt text for cover images should be "View on Amazon.com." Test in
both prog and Bootstrap.

NOTE: You'll need one to be overdue in order to trigger both
      cases. Examine the HTML for both tabs in the person details
      page.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
This fixed the cover alt texts for Amazon covers in the patron
account for checkouts, overdues and holds.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>