Bug 12325 - Marc and Card view in Z39.50 Results doesn't work properly in IE 8
This patch removes an unnecessary and problematic "return true" which
causes conflicts with Greybox when clicking on "MARC" or "Card" in
the search results of a Z39.50 search. The result is that the user is
taken to the showmarc.pl page and cannot return to the results, which
makes Z39.50 searching pretty much unusable for IE 8 users.

This patch also cleans up the "inline links" display for the separate
Z39.50 search pages. The 'biblio' version now has inline links for LCCN,
and the 'authority' version now prevents inline links for the MARC
and Import cells.

_TEST PLAN_

1) Open Internet Explorer 8 (if you don't have IE 8, use whatever Internet
Explorer version you do have. You can use F12 developer tools to emulate it.)
2) Go to Cataloguing > New from Z39.50
3) Do a search that will get results

(N.B. if you're using IE9+, hit F12 in the pop-up and change to IE 8 mode)

4) Click on "MARC" or "Card".
5) Note that Greybox tries to open a pop-over, but instead follows the
link to showmarc.pl. (If you're on IE 8, you can't use "Back" to go
back to the search results. If you're using an emulated version, you
can.)

(If you're testing the inline links, click on the LCCN cell for a biblio
search (not that no inline links appear), or MARC and Import cells when
doing an authority search (not that inline links do appear). Look at
bug 10096 for authority z39.50 server details.)

6) Apply the patch

7) Repeat steps 1-5
8) Note that Greybox now appears correctly when clicking "MARC" and
"Card"

9) Note also that inline links now appear for LCCN cell for biblio
searches, and inline links no longer appear for MARC and Import
cells when doing authority searches.

10) Profit

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Tested for regressions in Firefox and Chromium, no problems found.
Passes tests and QA script.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>